requirements.txt
Django==1.7.5
coverage==3.7.1
django-cors-headers==1.0.0
django-debug-toolbar==1.2.2
djangorestframework==3.0.5
requests==2.5.3
sqlparse==0.1.14
wsgiref==0.1.2